clean the code

+++ src/myApp/user.go
@@ -1,7 +1,7 @@
 package main
 
 type User struct {
-	Id    int64
+	Id    int32
 	Name  string `xorm:"varchar(25) not null unique 'usr_name'"`
 	Email string
 }
